refactor categories fragment into separate file
